spline #1011 added tests and parsing
